config restructured
