+++ hw12_13_14_15_16_calendar/internal/config/calendar/calendar_test.go
@@ -1,4 +1,4 @@
-package config
+package calendar
 
 import (
 	"reflect"
